First possibly workable version of the fms and secure token auth handlers

+++ src/org/red5/server/net/rtmp/RTMPConnection.java
@@ -77,9 +77,9 @@ import org.slf4j.LoggerFactory;
  * (AMF0/AMF3), connection state (is alive, last ping time and ping result) and
  * session.
  */
-public abstract class RTMPConnection extends BaseConnection implements
-		IStreamCapableConnection, IServiceCapableConnection {
-	
+public abstract class RTMPConnection extends BaseConnection implements IStreamCapableConnection,
+		IServiceCapableConnection {
+
 	/**
 	 * Logger
 	 */
@@ -242,7 +242,7 @@ public abstract class RTMPConnection extends BaseConnection implements
 			getReadLock().unlock();
 		}
 	}
-	
+
 	public byte getStateCode() {
 		getReadLock().lock();
 		try {
@@ -251,7 +251,7 @@ public abstract class RTMPConnection extends BaseConnection implements
 			getReadLock().unlock();
 		}
 	}
-	
+
 	public void setStateCode(byte code) {
 		getWriteLock().lock();
 		try {
@@ -270,7 +270,7 @@ public abstract class RTMPConnection extends BaseConnection implements
 			getWriteLock().unlock();
 		}
 	}
-	
+
 	@Override
 	public boolean connect(IScope newScope, Object[] params) {
 		log.debug("Connect scope: {}", newScope);
@@ -284,7 +284,8 @@ public abstract class RTMPConnection extends BaseConnection implements
 					// the bandwidth system-wide.
 					IScope s = getScope();
 					if (s != null && s.getContext() != null) {
-						IBWControlService bwController = (IBWControlService) s.getContext().getBean(IBWControlService.KEY);
+						IBWControlService bwController = (IBWControlService) s.getContext().getBean(
+								IBWControlService.KEY);
 						bwContext = bwController.registerBWControllable(this);
 					}
 					unscheduleWaitForHandshakeJob();
@@ -321,8 +322,7 @@ public abstract class RTMPConnection extends BaseConnection implements
 	 * @param sessionId Connection session id
 	 * @param params Params passed from client
 	 */
-	public void setup(String host, String path, String sessionId,
-			Map<String, Object> params) {
+	public void setup(String host, String path, String sessionId, Map<String, Object> params) {
 		this.host = host;
 		this.path = path;
 		this.sessionId = sessionId;
@@ -439,7 +439,7 @@ public abstract class RTMPConnection extends BaseConnection implements
 	/** {@inheritDoc} */
 	public IClientBroadcastStream newBroadcastStream(int streamId) {
 		getReadLock().lock();
-		try{
+		try {
 			if (!reservedStreams.get(streamId - 1)) {
 				// StreamId has not been reserved before
 				return null;
@@ -474,8 +474,7 @@ public abstract class RTMPConnection extends BaseConnection implements
 	/**
 	 * {@inheritDoc} To be implemented.
 	 */
-	public ISingleItemSubscriberStream newSingleItemSubscriberStream(
-			int streamId) {
+	public ISingleItemSubscriberStream newSingleItemSubscriberStream(int streamId) {
 		// TODO implement it
 		return null;
 	}
@@ -500,8 +499,8 @@ public abstract class RTMPConnection extends BaseConnection implements
 		 * Picking up the PlaylistSubscriberStream defined as a Spring
 		 * prototype in red5-common.xml
 		 */
-		PlaylistSubscriberStream pss = (PlaylistSubscriberStream) scope
-				.getContext().getBean("playlistSubscriberStream");
+		PlaylistSubscriberStream pss = (PlaylistSubscriberStream) scope.getContext()
+				.getBean("playlistSubscriberStream");
 		Integer buffer = streamBuffers.get(streamId - 1);
 		if (buffer != null) {
 			pss.setClientBufferDuration(buffer);
@@ -609,11 +608,10 @@ public abstract class RTMPConnection extends BaseConnection implements
 			getWriteLock().unlock();
 		}
 		Red5.setConnectionLocal(this);
-		IStreamService streamService = (IStreamService) getScopeService(scope,
-				IStreamService.class, StreamService.class);
+		IStreamService streamService = (IStreamService) getScopeService(scope, IStreamService.class,
+				StreamService.class);
 		if (streamService != null) {
-			for (Map.Entry<Integer, IClientStream> entry : streams
-					.entrySet()) {
+			for (Map.Entry<Integer, IClientStream> entry : streams.entrySet()) {
 				IClientStream stream = entry.getValue();
 				if (stream != null) {
 					log.debug("Closing stream: {}", stream.getStreamId());
@@ -627,10 +625,9 @@ public abstract class RTMPConnection extends BaseConnection implements
 
 		getWriteLock().lock();
 		try {
-			if (bwContext != null && getScope() != null
-					&& getScope().getContext() != null) {
-				IBWControlService bwController = (IBWControlService) getScope()
-						.getContext().getBean(IBWControlService.KEY);
+			if (bwContext != null && getScope() != null && getScope().getContext() != null) {
+				IBWControlService bwController = (IBWControlService) getScope().getContext().getBean(
+						IBWControlService.KEY);
 				bwController.unregisterBWControllable(bwContext);
 				bwContext = null;
 			}
@@ -715,11 +712,8 @@ public abstract class RTMPConnection extends BaseConnection implements
 	public void receivedBytesRead(int bytes) {
 		getWriteLock().lock();
 		try {
-			log
-				.debug(
-						"Client received {} bytes, written {} bytes, {} messages pending",
-						new Object[] { bytes, getWrittenBytes(),
-								getPendingMessages() });
+			log.debug("Client received {} bytes, written {} bytes, {} messages pending", new Object[] { bytes,
+					getWrittenBytes(), getPendingMessages() });
 			clientBytesRead = bytes;
 		} finally {
 			getWriteLock().unlock();
@@ -802,7 +796,7 @@ public abstract class RTMPConnection extends BaseConnection implements
 
 	/** {@inheritDoc} */
 	public void notify(IServiceCall call) {
-		notify(call, (byte) 3);
+		notify(call, 3);
 	}
 
 	/** {@inheritDoc} */
@@ -845,20 +839,18 @@ public abstract class RTMPConnection extends BaseConnection implements
 			return;
 		}
 
-		IConnectionBWConfig connectionBWConfig = (IConnectionBWConfig) config; 
+		IConnectionBWConfig connectionBWConfig = (IConnectionBWConfig) config;
 
 		// Notify client about new bandwidth settings (in bytes per second)
 		if (connectionBWConfig.getDownstreamBandwidth() > 0) {
-			ServerBW serverBW = new ServerBW((int) connectionBWConfig
-					.getDownstreamBandwidth() / 8);
+			ServerBW serverBW = new ServerBW((int) connectionBWConfig.getDownstreamBandwidth() / 8);
 			getChannel(2).write(serverBW);
 		}
 		if (connectionBWConfig.getUpstreamBandwidth() > 0) {
-			ClientBW clientBW = new ClientBW((int) connectionBWConfig
-					.getUpstreamBandwidth() / 8, (byte) 0);
+			ClientBW clientBW = new ClientBW((int) connectionBWConfig.getUpstreamBandwidth() / 8, (byte) 0);
 			getChannel(2).write(clientBW);
 		}
-		
+
 		getWriteLock().lock();
 		try {
 			this.bwConfig = connectionBWConfig;
@@ -870,8 +862,8 @@ public abstract class RTMPConnection extends BaseConnection implements
 			}
 
 			if (bwContext != null) {
-				IBWControlService bwController = (IBWControlService) getScope()
-						.getContext().getBean(IBWControlService.KEY);
+				IBWControlService bwController = (IBWControlService) getScope().getContext().getBean(
+						IBWControlService.KEY);
 				bwController.updateBWConfigure(bwContext);
 			}
 		} finally {
@@ -978,16 +970,15 @@ public abstract class RTMPConnection extends BaseConnection implements
 	@Override
 	public long getPendingVideoMessages(int streamId) {
 		AtomicInteger count = pendingVideos.get(streamId);
-		long result = (count != null ? count.intValue() - getUsedStreamCount()
-				: 0);
+		long result = (count != null ? count.intValue() - getUsedStreamCount() : 0);
 		return (result > 0 ? result : 0);
 	}
 
 	/** {@inheritDoc} */
 	public void ping() {
 		long newPingTime = System.currentTimeMillis();
-		log.debug("Pinging client with id {} at {}, last ping sent at {}", 
-				new Object[]{ getId(), newPingTime, lastPingSent.get() });
+		log.debug("Pinging client with id {} at {}, last ping sent at {}", new Object[] { getId(), newPingTime,
+				lastPingSent.get() });
 		if (lastPingSent.get() == 0) {
 			lastPongReceived.set(newPingTime);
 		}
@@ -1008,8 +999,8 @@ public abstract class RTMPConnection extends BaseConnection implements
 	protected void pingReceived(Ping pong) {
 		long now = System.currentTimeMillis();
 		long previousReceived = (int) (lastPingSent.get() & 0xffffffff);
-		log.debug("Pong from client id {} at {} with value {}, previous received at {}",
-				new Object[]{ getId(), now , pong.getValue2(), previousReceived });
+		log.debug("Pong from client id {} at {} with value {}, previous received at {}", new Object[] { getId(), now,
+				pong.getValue2(), previousReceived });
 		if (pong.getValue2() == previousReceived) {
 			lastPingTime.set((int) (now & 0xffffffff) - pong.getValue2());
 		}
@@ -1089,11 +1080,9 @@ public abstract class RTMPConnection extends BaseConnection implements
 	public String toString() {
 		// http://java.sun.com/j2se/1.5.0/docs/api/java/lang/String.html#format(
 		// java.lang.String,%20java.lang.Object...)
-		Object[] args = new Object[] { getClass().getSimpleName(),
-				getRemoteAddress(), getRemotePort(), getHost(), getReadBytes(),
-				getWrittenBytes() };
-		return String.format(
-				"%1$s from %2$s : %3$s to %4$s (in: %5$s out %6$s )", args);
+		Object[] args = new Object[] { getClass().getSimpleName(), getRemoteAddress(), getRemotePort(), getHost(),
+				getReadBytes(), getWrittenBytes() };
+		return String.format("%1$s from %2$s : %3$s to %4$s (in: %5$s out %6$s )", args);
 	}
 
 	/**
@@ -1159,8 +1148,9 @@ public abstract class RTMPConnection extends BaseConnection implements
 	private class KeepAliveJob implements IScheduledJob {
 
 		private final AtomicLong lastBytesRead = new AtomicLong(0);
+
 		private volatile long lastBytesReadTime = 0;
-		
+
 		/** {@inheritDoc} */
 		public void execute(ISchedulingService service) {
 			long thisRead = getReadBytes();
@@ -1174,8 +1164,7 @@ public abstract class RTMPConnection extends BaseConnection implements
 			}
 			// Client didn't send response to ping command 
 			// and didn't sent data for too long, disconnect
-			if (lastPongReceived.get() > 0
-					&& (lastPingSent.get() - lastPongReceived.get() > maxInactivity)
+			if (lastPongReceived.get() > 0 && (lastPingSent.get() - lastPongReceived.get() > maxInactivity)
 					&& !(System.currentTimeMillis() - lastBytesReadTime < maxInactivity)) {
 
 				getWriteLock().lock();
@@ -1192,13 +1181,13 @@ public abstract class RTMPConnection extends BaseConnection implements
 				} finally {
 					getWriteLock().unlock();
 				}
-				log.warn("Closing {}, with id {}, due to too much inactivity ({}ms), last ping sent {}ms ago", 
-						new Object[] { RTMPConnection.this, getId(), 
-							(lastPingSent.get() - lastPongReceived.get()),
-							(System.currentTimeMillis() - lastPingSent.get())});
+				log.warn("Closing {}, with id {}, due to too much inactivity ({}ms), last ping sent {}ms ago",
+						new Object[] { RTMPConnection.this, getId(), (lastPingSent.get() - lastPongReceived.get()),
+								(System.currentTimeMillis() - lastPingSent.get()) });
 				// Add the following line to (hopefully) deal with a very common support request
 				// on the Red5 list
-				log.warn("This often happens if YOUR Red5 application generated an exception on start-up.  Check earlier in the log for that exception first!");
+				log
+						.warn("This often happens if YOUR Red5 application generated an exception on start-up.  Check earlier in the log for that exception first!");
 				onInactive();
 				return;
 			}
